Feat/ Retrieve job results as dict instead of string
Job results are now sent in a dictionnary instead of in a string. Update the job class accordingly.

A new API endpoint has been defined to retrieve all job results for a given batch. Use that endpoint to retrieve job results when asked by the user instead of making a request for each job separately